Leaf area index (LAI)_Back Propagation
Leaf area index (LAI) is an important parameter of vegetation structure and the key parameter of biophysical change, which can provide dynamic information of vegetation growth. In order to study the agricultural development in a region of Ukraine, the LAI sample data of 100 local ground measurements were obtained from foreign websites. According to the coordinates of these sample points, the reflectance data of landsat8 satellite from 1 to 7 bands are extracted.